+++ SpiceQL/include/io.h
@@ -6,6 +6,7 @@
  *
  **/
 
+#include <cstddef>
 #include <optional>
 #include <string>
 #include <vector>