+++ SpiceQL/src/api.cpp
@@ -33,11 +33,16 @@ namespace SpiceQL {
     vector<string> default_KernelQualities = {"smithed", "reconstructed"};
 
     json aliasMap = {
+      {"A15_METRIC", "apollo15_metric"},
+      {"APOLLO15_APOLLO PANORAMIC CAMERA", "apollo15_panoramic"},
       {"AMICA", "amica"},
       {"CHANDRAYAAN-1_M3", "m3"},
       {"CHANDRAYAAN-1_MRFFR", "mrffr"},
       {"CASSINI_ISS_NAC", "cassini"},
       {"CASSINI_ISS_WAC", "cassini"},
+      {"CASSINI_VIMS_V", "cassini"},
+      {"Cassini-Huygens", "cassini"},
+      {"CONTEXT CAMERA", "ctx"},
       {"DAWN_FC2_FILTER_1", "fc2"},
       {"DAWN_FC2_FILTER_2", "fc2"},
       {"DAWN_FC2_FILTER_3", "fc2"},
@@ -46,6 +51,7 @@ namespace SpiceQL {
       {"DAWN_FC2_FILTER_6", "fc2"},
       {"DAWN_FC2_FILTER_7", "fc2"},
       {"DAWN_FC2_FILTER_8", "fc2"},
+      {"DAWN_VIR_IR", "vir"},
       {"GLL_SSI_PLATFORM", "galileo"},
       {"HAYABUSA_AMICA", "amica"},
       {"HAYABUSA_NIRS", "nirs"},
@@ -59,9 +65,12 @@ namespace SpiceQL {
       {"LRO_MINIRF", "minirf"},
       {"M10_VIDICON_A", "m10_vidicon_a"},
       {"M10_VIDICON_B", "m10_vidicon_b"},
+      {"VIDICON_A", "m10_vidicon_a"},
       {"MARS", "mro"},
       {"MSGR_MDIS_WAC", "mdis"},
       {"MSGR_MDIS_NAC", "mdis"},
+      {"MERCURY DUAL IMAGING SYSTEM NARROW ANGLE CAMERA", "mdis"},
+      {"MEX_HRSC_HEAD", "hrsc"},
       {"MEX_HRSC_SRC", "src"},
       {"MEX_HRSC_IR", "hrsc"},
       {"MGS_MOC_NA", "mgs"},
@@ -69,14 +78,22 @@ namespace SpiceQL {
       {"MGS_MOC_WA_BLUE", "mgs"},
       {"MRO_MARCI_VIS", "marci"},
       {"MRO_MARCI_UV", "marci"},
+      {"MRO_MARCI_BASE", "marci"},
       {"MRO_CTX", "ctx"},
       {"MRO_HIRISE", "hirise"},
+      {"MRO_HIRISE_LOOK_DIRECTION", "hirise"},
       {"MRO_CRISM_VNIR", "crism"},
-      {"NEAR EARTH ASTEROID RENDEZVOUS", ""},
+      {"NEAR EARTH ASTEROID RENDEZVOUS", "near"},
+      {"NEAR_MSI", "msi"},
       {"NH_LORRI", "lorri"},
+      {"NH_LORRI_1X1", "lorri"},
       {"NH_RALPH_LEISA", "leisa"},
       {"NH_MVIC", "mvic_tdi"},
+      {"ISIS_NH_RALPH_LEISA", "leisa"},
       {"ISIS_NH_RALPH_MVIC_METHANE", "mvic_framing"},
+      {"ISIS_NH_RALPH_MVIC_FT", "mvic_framing"},
+      {"M01_THEMIS_IR", "odyssey"},
+      {"M01_THEMIS_VIS", "odyssey"},
       {"THEMIS_IR", "odyssey"},
       {"THEMIS_VIS", "odyssey"},
       {"LISM_MI-VIS1", "kaguya"},
@@ -101,16 +118,27 @@ namespace SpiceQL {
       {"LISM_TC1_SDH", "kaguya"},
       {"LISM_TC1_STH", "kaguya"},
       {"LISM_TC1_SSH", "kaguya"},
+      {"SELENE", "kaguya"},
+      {"KAGUYA", "kaguya"},
+      {"SELENE MAIN ORBITER", "kaguya"},
       {"LO1_HIGH_RESOLUTION_CAMERA", "lo"},
       {"LO2_HIGH_RESOLUTION_CAMERA", "lo"},
       {"LO3_HIGH_RESOLUTION_CAMERA", "lo"},
       {"LO4_HIGH_RESOLUTION_CAMERA", "lo"},
       {"LO5_HIGH_RESOLUTION_CAMERA", "lo"},
+      {"LO3_MED_RESOLUTION_CAMERA", "lo"},
       {"NEPTUNE", "voyager1"}, 
       {"SATURN", "voyager1"},
       {"TGO_CASSIS", "cassis"},
+      {"TGO_CASSIS_CRU", "cassis"},
       {"VIKING ORBITER 1", "viking1"},
+      {"VIKING_ORBITER_1", "viking1"},
       {"VIKING ORBITER 2", "viking2"},
+      {"VIKING_ORBITER_2", "viking2"},
+      {"VO2_VISB", "viking2"},
+      {"VO2_VISA", "viking2"},
+      {"VO1_VISA", "viking1"},
+      {"VO1_VISB", "viking1"},
       {"VG1_ISSNA", "voyager1"},
       {"VG1_ISSWA", "voyager1"},
       {"VG2_ISSNA", "voyager2"},
@@ -120,8 +148,16 @@ namespace SpiceQL {
       {"High Resolution Camera", "clementine1"},
       {"Long Wave Infrared Camera", "clementine1"},
       {"Visual and Infrared Spectrometer", "vir"},
+      {"CLEM_HIRES", "clementine1"},
+      {"CLEM_UVVIS_A", "uvvis"},
+      {"CLEM_LWIR", "clementine1"},
+      {"CLEM_NIR", "nir"},
       {"CH2", "chandrayaan2"},
-      {"CH-2", "chandrayaan2"}
+      {"CH-2", "chandrayaan2"},
+      {"CH2_TMC_NADIR", "tmc2"},
+      {"CH2_OHRC", "ohrc"},
+      {"ROS_VIRTIS-M_IR", "virtis"},
+      {"MSL_MASTCAM_LEFT", "msl"}
     };
     
     /**
@@ -814,7 +850,7 @@ namespace SpiceQL {
         json kernelsToLoad = {};
 
         if (mission != "" && searchKernels){
-            kernelsToLoad = Inventory::search_for_kernelset(mission, {"fk"}, default_StartTime, default_StopTime, default_KernelQualities, default_KernelQualities, fullKernelPath, limitCk, limitSpk);
+            kernelsToLoad = Inventory::search_for_kernelset(mission, {"fk", "ik", "iak"}, default_StartTime, default_StopTime, default_KernelQualities, default_KernelQualities, fullKernelPath, limitCk, limitSpk);
         }
         if (!kernelList.empty()) {
             json regexk = Inventory::search_for_kernelset_from_regex(kernelList, fullKernelPath);